Flashcard Lecturer refactoring
